Update UTF8 Encoding Changes
Byte code is returned as a hex number ranging from 0-255. The UTF8 conversion halves the size by turning the hex number (stored as 2 chars) into a UTF8 char. Previously this was shifted from UTF8 0-255 to UTF8 1-256. Unfortunately this caused issues when using copy and paste and reading files. This has been shifted to 183-438. This will help with making the UTF8 ByteCode Encoding more friendly to use as copy/paste issues will be fixed.

There is a new settings object in "JStoJSC" that contains an option to use raw bytes for UTF8 aswell as arrays. This will force the byte range to be 0-255. A new arg has been added to array conversion to specifiy if the array is already in a "raw" byte form or if the UTF8 chars are in a "raw" byte form.

The minified version has been updated as well as the github.io version.